Compile under gcc version 4.1.1.

+++ lib/misc/BootLibrary.c
@@ -36,7 +36,7 @@ void * memcpy(void * to, const void * from, size_t n)
 }
 
 
-size_t strlen(const char * s)
+int strlen(const char * s)
 {
 	int d0;
 	register int __res;
@@ -99,7 +99,7 @@ char * strcpy(char * dest,const char *src)
 	return dest;
 }
 
-char * strncpy(char * dest,const char *src,size_t count)
+char * strncpy(char * dest,const char *src,int count)
 {
 	int d0, d1, d2, d3;
 	__asm__ __volatile__(
